requirements.txt
beautifulsoup4
lxml
requests
markdown
latlon3